Final commit V1
Final commit

Changed a few things. gave two player and one player options
with a menu. theres a restart button now and the cpu goes a bit slower everytime it hits the ball to compensate for its perfection

Will come back to this project. Moving on to flappy bird. Will come back and change one or two things every now and then